Replace `postgres` with `postgrespool`
This should increase relaibility and possibly performance, at the
expense of more database connections.
